+++ unlock-app/src/components/interface/locks/Settings/elements/SettingTerms.tsx
@@ -3,6 +3,7 @@ import { ReactNode } from 'react'
 import { CancellationForm } from '../forms/CancellationForm'
 import { UpdateDurationForm } from '../forms/UpdateDurationForm'
 import { UpdateQuantityForm } from '../forms/UpdateQuantityForm'
+import { UpdateTransferFee } from '../forms/UpdateTransferFee'
 import { SettingCard } from './SettingCard'
 
 interface SettingTermsProps {
@@ -29,7 +30,7 @@ export const SettingTerms = ({
   const settings: SettingProps[] = [
     {
       label: 'Duration',
-      description: 'Set up how long each membership lasts.',
+      description: 'Set up how long each membership lasts. ',
       children: (
         <UpdateDurationForm
           lockAddress={lockAddress}
@@ -43,7 +44,7 @@ export const SettingTerms = ({
     {
       label: 'Quantity',
       description:
-        'The maximum number of memberships that can be sold. Note: There is no limit to the number of memberships that can be airdropped by a lock manager or key granter.',
+        'The maximum number of memberships that can be sold from your contract. Note: there is no limit to the number of memberships that can be airdropped by a lock manager or key granter.',
       children: (
         <UpdateQuantityForm
           lockAddress={lockAddress}
@@ -55,8 +56,15 @@ export const SettingTerms = ({
     },
     {
       label: 'Transfer',
-      description: 'Allow members to transfer membership from one to others.',
-      children: null,
+      description: 'Allow members to transfer memberships.',
+      children: (
+        <UpdateTransferFee
+          lockAddress={lockAddress}
+          network={network}
+          isManager={isManager}
+          disabled={!isManager}
+        />
+      ),
     },
     {
       label: 'Cancellation',